move again, consolidate
