+++ Cargo.toml
@@ -1,6 +1,38 @@
-[workspace]
+[package]
+name = "holochain-runner"
+version = "0.0.31"
+description = "Gives a binary providing wrapped Holochain Conductor with Status Update events, and a good SIGTERM kill switch"
+authors = ["Connor Turland <connor@sprillow.com>"]
+license = "CAL-1.0"
+repository = "https://github.com/Sprillow/holochain-runner"
+edition = "2018"
 
-members = [
-    "crates/holochain-runner",
-    "crates/embedded-holochain-runner",
-]
+[dependencies]
+# these are just here as patches for
+#  a version mismatch with wasmer
+# can be removed in the future
+wasmer-engine-universal = "=2.0.0"
+wasmer-engine-dylib = "=2.0.0"
+wasmer-engine = "=2.0.0"
+wasmer-compiler-cranelift = "=2.0.0"
+wasmer-compiler = "=2.0.0"
+wasmer-vm = "=2.0.0"
+wasmer-derive = "=2.0.0"
+wasmer-types = "=2.0.0"
+
+structopt = "0.3.11"
+base64 = "0.13.0"
+serde = "1.0.123"
+futures = "0.3.1"
+human-panic = "1.0.3"
+tracing = "0.1.26"
+tokio = { version = "1.3", features = ["full"] }
+holochain_util = "0.0.4"
+holochain_p2p = "0.0.15"
+holochain_types = "0.0.15"
+holochain_zome_types = "0.0.17"
+holochain_keystore = "0.0.15"
+holochain = "0.0.115"
+hdk = "0.0.115"
+# termination feature enables handling of SIGTERM events as well on unix
+ctrlc = { version = "3.2.0", features = ["termination"] }
