--- crates/embedded-holochain-runner/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "embedded_holochain_runner"
-description = "A library that makes it VERY easy to run Holochain as a library, from your own binary, with great defaults"
-license = "CAL-1.0"
-version = "0.0.7"
-authors = ["Connor Turland <connor@sprillow.com>"]
-edition = "2018"
-
-[lib]
-
-[dependencies]
-# these are just here as patches for
-#  a version mismatch with wasmer
-# can be removed in the future
-wasmer-engine-universal = "=2.0.0"
-wasmer-engine-dylib = "=2.0.0"
-wasmer-engine = "=2.0.0"
-wasmer-compiler-cranelift = "=2.0.0"
-wasmer-compiler = "=2.0.0"
-wasmer-vm = "=2.0.0"
-wasmer-derive = "=2.0.0"
-wasmer-types = "=2.0.0"
-
-base64 = "0.13.0"
-serde = "1.0.123"
-futures = "0.3.1"
-human-panic = "1.0.3"
-tracing = "0.1.26"
-tokio = { version = "1.3", features = ["full"] }
-holochain_util = "0.0.4"
-holochain_p2p = "0.0.15"
-holochain_types = "0.0.15"
-holochain_zome_types = "0.0.17"
-holochain_keystore = "0.0.15"
-holochain = "0.0.115"
-hdk = "0.0.115"
-ctrlc = { version = "3.2.0", features = ["termination"] }
-
-# [patch.crates-io]
-# rkv = { git = "https://github.com/holochain/rkv.git", branch = "master" }
-# lmdb-rkv = { git = "https://github.com/holochain/lmdb-rs.git" }
