--- crates/embedded-holochain-runner/src/lib.rs
@@ -1,8 +1,0 @@
-mod embedded_runner;
-mod config;
-mod install_enable;
-mod emit;
-mod generate_key;
-
-pub use embedded_runner::*;
-pub use emit::StateSignal;
