+++ dry-run.sh
@@ -3,4 +3,4 @@
 set -euo pipefail
 
 # cargo run --manifest-path crates/holochain-runner/Cargo.toml --release -- --keystore-path test/data/keystore test/profiles.dna test/data/databases
-cargo run --manifest-path crates/holochain-runner/Cargo.toml -- --keystore-path test/data/keystore test/profiles.dna test/data/databases
+cargo run --release -- --keystore-path test/data/keystore test/profiles.dna test/data/databases
