+++ src/emit.rs
